Create Carrot Cake Cookies recipe page
Created carrot_cake_cookies.html page with the same format as the
pumpkin_snickerdoodles.html page to stay consistent. Also cleaned up
some of the missing periods in the steps for the pumpkin_snickerdoodles
recipe. Resized images to 500 height to bring it closer to the scale of
the rest of the page.